pyproject.toml
[tool.poetry]
name = "enterprise-fastapi-template"
version = "0.1.0"
description = "Enterprise FastAPI starter template"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111"
uvicorn = {extras = ["standard"], version = "^0.29"}
pydantic = "^2.7"
pydantic-settings = "^2.2"
python-jose = {extras = ["cryptography"], version = "^3.3"}
passlib = {extras = ["bcrypt"], version = "^1.7"}
httpx = "^0.27"
sqlalchemy = "^2.0"
alembic = "^1.13"
asyncpg = "^0.29"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-asyncio = "^0.23"
pytest-cov = "^5.0"
ruff = "^0.4"
mypy = "^1.10"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP"]

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
